+++ inst/shinydemos/sortable_bookmarking.R
@@ -11,7 +11,7 @@ ui <- function(request) {fluidPage(
 server <- function(input, output) {
   output$order <- renderPrint({
     cat("Rows order:\n")
-    input$tbl_index
+    input$tbl_order
   })
   output$tbl <- renderTable(head(mtcars), rownames = TRUE)